IGNITE-24921 Make ReplicaStateManager a top-level class
