+++ modules/replicator/src/main/java/org/apache/ignite/internal/replicator/ReplicaManager.java
@@ -24,7 +24,6 @@ import static java.util.concurrent.CompletableFuture.supplyAsync;
 import static java.util.stream.Collectors.toSet;
 import static org.apache.ignite.internal.failure.FailureType.CRITICAL_ERROR;
 import static org.apache.ignite.internal.lang.IgniteStringFormatter.format;
-import static org.apache.ignite.internal.raft.PeersAndLearners.fromAssignments;
 import static org.apache.ignite.internal.replicator.LocalReplicaEvent.AFTER_REPLICA_STARTED;
 import static org.apache.ignite.internal.replicator.LocalReplicaEvent.BEFORE_REPLICA_STOPPED;
 import static org.apache.ignite.internal.replicator.message.ReplicaMessageUtils.toTablePartitionIdMessage;
@@ -33,10 +32,8 @@ import static org.apache.ignite.internal.thread.ThreadOperation.STORAGE_READ;
 import static org.apache.ignite.internal.thread.ThreadOperation.STORAGE_WRITE;
 import static org.apache.ignite.internal.thread.ThreadOperation.TX_STATE_STORAGE_ACCESS;
 import static org.apache.ignite.internal.util.CompletableFutures.allOf;
-import static org.apache.ignite.internal.util.CompletableFutures.falseCompletedFuture;
 import static org.apache.ignite.internal.util.CompletableFutures.isCompletedSuccessfully;
 import static org.apache.ignite.internal.util.CompletableFutures.nullCompletedFuture;
-import static org.apache.ignite.internal.util.CompletableFutures.trueCompletedFuture;
 import static org.apache.ignite.internal.util.ExceptionUtils.hasCauseOrSuppressed;
 import static org.apache.ignite.internal.util.ExceptionUtils.unwrapCause;
 import static org.apache.ignite.internal.util.IgniteUtils.shouldSwitchToRequestsExecutor;
@@ -46,7 +43,6 @@ import static org.apache.ignite.lang.ErrorGroups.Common.INTERNAL_ERR;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
 import java.util.UUID;
@@ -85,8 +81,6 @@ import org.apache.ignite.internal.network.NetworkMessage;
 import org.apache.ignite.internal.network.NetworkMessageHandler;
 import org.apache.ignite.internal.partitiondistribution.Assignments;
 import org.apache.ignite.internal.placementdriver.PlacementDriver;
-import org.apache.ignite.internal.placementdriver.event.PrimaryReplicaEvent;
-import org.apache.ignite.internal.placementdriver.event.PrimaryReplicaEventParameters;
 import org.apache.ignite.internal.placementdriver.message.PlacementDriverMessageGroup;
 import org.apache.ignite.internal.placementdriver.message.PlacementDriverMessagesFactory;
 import org.apache.ignite.internal.placementdriver.message.PlacementDriverReplicaMessage;
@@ -133,7 +127,6 @@ import org.apache.ignite.lang.IgniteException;
 import org.apache.ignite.network.ClusterNode;
 import org.jetbrains.annotations.Nullable;
 import org.jetbrains.annotations.TestOnly;
-import org.jetbrains.annotations.VisibleForTesting;
 
 /**
  * Replica manager maintains {@link Replica} instances on an Ignite node.
@@ -181,6 +174,8 @@ public class ReplicaManager extends AbstractEventProducer<LocalReplicaEvent, Loc
 
     /** Creator for {@link org.apache.ignite.internal.raft.storage.LogStorageFactory} for volatile tables. */
     private final LogStorageFactoryCreator volatileLogStorageFactoryCreator;
+
+    private final Executor replicaStartStopExecutor;
 
     /** Raft command marshaller for raft server endpoints starting. */
     private final Marshaller raftCommandsMarshaller;
@@ -283,8 +278,11 @@ public class ReplicaManager extends AbstractEventProducer<LocalReplicaEvent, Loc
         this.raftManager = raftManager;
         this.partitionRaftConfigurer = partitionRaftConfigurer;
         this.getPendingAssignmentsSupplier = getPendingAssignmentsSupplier;
+        this.replicaStartStopExecutor = replicaStartStopExecutor;
+
         this.replicaStateManager = new ReplicaStateManager(
                 replicaStartStopExecutor,
+                requestsExecutor,
                 clockService,
                 placementDriver,
                 this
@@ -517,7 +515,7 @@ public class ReplicaManager extends AbstractEventProducer<LocalReplicaEvent, Loc
      * @param redirectNodeId Node consistent id to redirect.
      * @return Future that is completed when the lease is denied to prolong, containing the expiration time of this lease.
      */
-    private CompletableFuture<HybridTimestamp> stopLeaseProlongation(
+    CompletableFuture<HybridTimestamp> stopLeaseProlongation(
             ReplicationGroupId groupId,
             @Nullable String redirectNodeId
     ) {
@@ -905,7 +903,7 @@ public class ReplicaManager extends AbstractEventProducer<LocalReplicaEvent, Loc
                     }
 
                     return true;
-                }, replicaStateManager.replicaStartStopPool);
+                }, replicaStartStopExecutor);
     }
 
     /** {@inheritDoc} */
@@ -1139,7 +1137,6 @@ public class ReplicaManager extends AbstractEventProducer<LocalReplicaEvent, Loc
      * @return True if the replica is started.
      */
     @TestOnly
-    @VisibleForTesting
     @Deprecated
     public boolean isReplicaStarted(ReplicationGroupId replicaGrpId) {
         CompletableFuture<Replica> replicaFuture = replicas.get(replicaGrpId);
@@ -1148,9 +1145,9 @@ public class ReplicaManager extends AbstractEventProducer<LocalReplicaEvent, Loc
 
     /**
      * Can possibly start replica if it's not running or is stopping. Nothing happens if the replica is already running
-     * ({@link ReplicaState#ASSIGNED} or {@link ReplicaState#PRIMARY_ONLY}) and {@code forcedAssignments} is {@code null}.
-     * If the replica is {@link ReplicaState#ASSIGNED} and {@code forcedAssignments} is not {@code null} then peers will be
-     * reset to the given assignments. See {@link ReplicaState} for exact replica state transitions.
+     * ({@code ReplicaState#ASSIGNED} or {@code ReplicaState#PRIMARY_ONLY}) and {@code forcedAssignments} is {@code null}.
+     * If the replica is {@code ReplicaState#ASSIGNED} and {@code forcedAssignments} is not {@code null} then peers will be
+     * reset to the given assignments. See {@code ReplicaState} for exact replica state transitions.
      *
      * @param groupId Group id.
      * @param startOperation Replica start operation. Will be called if this method decides to start the replica.
@@ -1169,8 +1166,8 @@ public class ReplicaManager extends AbstractEventProducer<LocalReplicaEvent, Loc
      * Can possibly stop replica if it is running or starting, and is not a primary replica. Relies on the given reason. If
      * the reason is {@link WeakReplicaStopReason#EXCLUDED_FROM_ASSIGNMENTS} then the replica can be not stopped if it is still
      * a primary. If the reason is {@link WeakReplicaStopReason#PRIMARY_EXPIRED} then the replica is stopped only if its state
-     * is {@link ReplicaState#PRIMARY_ONLY}, because this assumes that it was excluded from assignments before.
-     * See {@link ReplicaState} for exact replica state transitions.
+     * is {@code ReplicaState#PRIMARY_ONLY}, because this assumes that it was excluded from assignments before.
+     * See {@code ReplicaState} for exact replica state transitions.
      *
      * @param groupId Group id.
      * @param reason Reason to stop replica.
@@ -1209,354 +1206,6 @@ public class ReplicaManager extends AbstractEventProducer<LocalReplicaEvent, Loc
                 .collect(toSet());
     }
 
-    private static class ReplicaStateManager {
-        private static final IgniteLogger LOG = Loggers.forClass(ReplicaStateManager.class);
-
-        final Map<ReplicationGroupId, ReplicaStateContext> replicaContexts = new ConcurrentHashMap<>();
-
-        final Executor replicaStartStopPool;
-
-        final ClockService clockService;
-
-        final PlacementDriver placementDriver;
-
-        final ReplicaManager replicaManager;
-
-        volatile UUID localNodeId;
-
-        private final IgniteSpinBusyLock busyLock = new IgniteSpinBusyLock();
-
-        ReplicaStateManager(
-                Executor replicaStartStopPool,
-                ClockService clockService,
-                PlacementDriver placementDriver,
-                ReplicaManager replicaManager
-        ) {
-            this.replicaStartStopPool = replicaStartStopPool;
-            this.clockService = clockService;
-            this.placementDriver = placementDriver;
-            this.replicaManager = replicaManager;
-        }
-
-        void start(UUID localNodeId) {
-            this.localNodeId = localNodeId;
-            placementDriver.listen(PrimaryReplicaEvent.PRIMARY_REPLICA_ELECTED, this::onPrimaryElected);
-            placementDriver.listen(PrimaryReplicaEvent.PRIMARY_REPLICA_EXPIRED, this::onPrimaryExpired);
-        }
-
-        void stop() {
-            busyLock.block();
-        }
-
-        private CompletableFuture<Boolean> onPrimaryElected(PrimaryReplicaEventParameters parameters) {
-            // Busy lock guarding because on node stop we shouldn't handle primary replica events anymore.
-            if (!busyLock.enterBusy()) {
-                return failedFuture(new NodeStoppingException());
-            }
-
-            try {
-                ReplicationGroupId replicationGroupId = parameters.groupId();
-                ReplicaStateContext context = getContext(replicationGroupId);
-
-                synchronized (context) {
-                    if (localNodeId.equals(parameters.leaseholderId())) {
-                        assert context.replicaState != ReplicaState.STOPPED
-                                : "Unexpected primary replica state STOPPED [groupId=" + replicationGroupId
-                                    + ", leaseStartTime=" + parameters.startTime() + ", reservedForPrimary=" + context.reservedForPrimary
-                                    + ", contextLeaseStartTime=" + context.leaseStartTime + "].";
-                    } else if (context.reservedForPrimary) {
-                        context.assertReservation(replicationGroupId, parameters.startTime());
-
-                        // Unreserve if another replica was elected as primary, only if its lease start time is greater,
-                        // otherwise it means that event is too late relatively to lease negotiation start and should be ignored.
-                        if (parameters.startTime().compareTo(context.leaseStartTime) > 0) {
-                            context.unreserve();
-
-                            if (context.replicaState == ReplicaState.PRIMARY_ONLY) {
-                                executeDeferredReplicaStop(replicationGroupId, context);
-                            }
-                        }
-                    }
-                }
-
-                return falseCompletedFuture();
-            } finally {
-                busyLock.leaveBusy();
-            }
-        }
-
-        private CompletableFuture<Boolean> onPrimaryExpired(PrimaryReplicaEventParameters parameters) {
-            // Busy lock guarding because on node stop we shouldn't handle primary replica events anymore.
-            if (!busyLock.enterBusy()) {
-                return failedFuture(new NodeStoppingException());
-            }
-
-            try {
-                if (localNodeId.equals(parameters.leaseholderId())) {
-                    ReplicaStateContext context = replicaContexts.get(parameters.groupId());
-
-                    if (context != null) {
-                        synchronized (context) {
-                            context.assertReservation(parameters.groupId(), parameters.startTime());
-
-                            // Unreserve if primary replica expired, only if its lease start time is equal to reservation time,
-                            // otherwise it means that event is too late relatively to lease negotiation start and should be ignored.
-                            if (parameters.startTime().equals(context.leaseStartTime)) {
-                                context.unreserve();
-
-                                if (context.replicaState == ReplicaState.RESTART_PLANNED) {
-                                    executeDeferredReplicaStop(parameters.groupId(), context);
-                                }
-                            }
-                        }
-                    }
-                }
-
-                return falseCompletedFuture();
-            } finally {
-                busyLock.leaveBusy();
-            }
-        }
-
-        ReplicaStateContext getContext(ReplicationGroupId groupId) {
-            return replicaContexts.computeIfAbsent(groupId,
-                    // Treat the absence in the map as STOPPED.
-                    k -> new ReplicaStateContext(ReplicaState.STOPPED, nullCompletedFuture()));
-        }
-
-        /**
-         * Can possibly start replica if it's not running or is stopping.
-         *
-         * @param groupId Group id.
-         * @param startOperation Replica start operation.
-         * @param forcedAssignments Assignments to reset forcibly, if needed. Assignments reset is only available when replica is started.
-         * @return Completable future, the result means whether the replica was started.
-         */
-        CompletableFuture<Boolean> weakStartReplica(
-                ReplicationGroupId groupId,
-                Supplier<CompletableFuture<Boolean>> startOperation,
-                @Nullable Assignments forcedAssignments
-        ) {
-            ReplicaStateContext context = getContext(groupId);
-
-            synchronized (context) {
-                ReplicaState state = context.replicaState;
-
-                LOG.debug("Weak replica start [grp={}, state={}, future={}].", groupId, state, context.previousOperationFuture);
-
-                if (state == ReplicaState.STOPPED || state == ReplicaState.STOPPING) {
-                    return startReplica(groupId, context, startOperation);
-                } else if (state == ReplicaState.ASSIGNED) {
-                    if (forcedAssignments != null) {
-                        assert forcedAssignments.force() :
-                                format("Unexpected assignments to force [assignments={}, groupId={}].", forcedAssignments, groupId);
-
-                        replicaManager.resetPeers(groupId, fromAssignments(forcedAssignments.nodes()));
-                    }
-
-                    // Telling the caller that the replica is started.
-                    return trueCompletedFuture();
-                } else if (state == ReplicaState.PRIMARY_ONLY) {
-                    context.replicaState = ReplicaState.ASSIGNED;
-
-                    LOG.debug("Weak replica start complete [state={}].", context.replicaState);
-
-                    return trueCompletedFuture();
-                } else if (state == ReplicaState.RESTART_PLANNED) {
-                    throw new AssertionError("Replica start cannot begin before stop on replica restart is completed [groupId="
-                            + groupId + "].");
-                } // else no-op.
-
-                throw new AssertionError("Replica start cannot begin while the replica is being started [groupId=" + groupId + "].");
-            }
-        }
-
-        private CompletableFuture<Boolean> startReplica(
-                ReplicationGroupId groupId,
-                ReplicaStateContext context,
-                Supplier<CompletableFuture<Boolean>> startOperation
-        ) {
-            context.replicaState = ReplicaState.STARTING;
-            context.previousOperationFuture = context.previousOperationFuture
-                    .handleAsync((v, e) -> startOperation.get(), replicaStartStopPool)
-                    .thenCompose(startOperationFuture -> startOperationFuture.thenApply(partitionStarted -> {
-                        synchronized (context) {
-                            if (partitionStarted) {
-                                context.replicaState = ReplicaState.ASSIGNED;
-                            } else {
-                                context.replicaState = ReplicaState.STOPPED;
-                                replicaContexts.remove(groupId);
-                            }
-                        }
-
-                        LOG.debug("Weak replica start complete [state={}, partitionStarted={}].", context.replicaState, partitionStarted);
-
-                        return partitionStarted;
-                    }))
-                    .exceptionally(e -> {
-                        LOG.error("Replica start failed [groupId={}]", e, groupId);
-
-                        throw new CompletionException(e);
-                    });
-
-            return context.previousOperationFuture;
-        }
-
-        /**
-         * Can possibly stop replica if it is running or starting, and is not a primary replica. Relies on the given reason. If
-         * the reason is {@link WeakReplicaStopReason#EXCLUDED_FROM_ASSIGNMENTS} then the replica can be not stopped if it is still
-         * a primary. If the reason is {@link WeakReplicaStopReason#PRIMARY_EXPIRED} then the replica is stopped only if its state
-         * is {@link ReplicaState#PRIMARY_ONLY}, because this assumes that it was excluded from assignments before.
-         *
-         * @param groupId Group id.
-         * @param reason Reason to stop replica.
-         * @param stopOperation Replica stop operation.
-         * @return Completable future, the result means whether the replica was stopped.
-         */
-        CompletableFuture<Void> weakStopReplica(
-                ReplicationGroupId groupId,
-                WeakReplicaStopReason reason,
-                Supplier<CompletableFuture<Void>> stopOperation
-        ) {
-            ReplicaStateContext context = getContext(groupId);
-
-            synchronized (context) {
-                ReplicaState state = context.replicaState;
-
-                LOG.debug("Weak replica stop [grpId={}, state={}, reason={}, reservedForPrimary={}, future={}].", groupId, state,
-                        reason, context.reservedForPrimary, context.previousOperationFuture);
-
-                if (reason == WeakReplicaStopReason.EXCLUDED_FROM_ASSIGNMENTS) {
-                    if (state == ReplicaState.ASSIGNED) {
-                        if (context.reservedForPrimary) {
-                            context.replicaState = ReplicaState.PRIMARY_ONLY;
-                            // Intentionally do not return future here: it can freeze the handling of assignment changes.
-                            planDeferredReplicaStop(groupId, context, null, stopOperation);
-                        } else {
-                            return stopReplica(groupId, context, stopOperation);
-                        }
-                    } else if (state == ReplicaState.STARTING) {
-                        return stopReplica(groupId, context, stopOperation);
-                    } else if (state == ReplicaState.STOPPED) {
-                        // We need to stop replica and destroy storages anyway, because they can be already created.
-                        return stopReplica(groupId, context, stopOperation);
-                    } // else: no-op.
-                } else if (reason == WeakReplicaStopReason.RESTART) {
-                    // Explicit restart: always stop.
-                    if (context.reservedForPrimary) {
-                        // If is primary, turning off the primary first.
-                        context.replicaState = ReplicaState.RESTART_PLANNED;
-                        return replicaManager.stopLeaseProlongation(groupId, null)
-                                .thenCompose(leaseExpirationTime ->
-                                        planDeferredReplicaStop(groupId, context, leaseExpirationTime, stopOperation)
-                                );
-                    } else {
-                        return stopReplica(groupId, context, stopOperation);
-                    }
-                } else {
-                    assert reason == WeakReplicaStopReason.PRIMARY_EXPIRED : "Unknown replica stop reason: " + reason;
-
-                    if (state == ReplicaState.PRIMARY_ONLY) {
-                        return stopReplica(groupId, context, stopOperation);
-                    } // else: no-op.
-                }
-                // State #RESTART_PLANNED is also no-op because replica will be stopped within deferred operation.
-
-                LOG.debug("Weak replica stop (sync part) complete [grpId={}, state={}].", groupId, context.replicaState);
-
-                return nullCompletedFuture();
-            }
-        }
-
-        private CompletableFuture<Void> stopReplica(
-                ReplicationGroupId groupId,
-                ReplicaStateContext context,
-                Supplier<CompletableFuture<Void>> stopOperation
-        ) {
-            context.replicaState = ReplicaState.STOPPING;
-            context.previousOperationFuture = context.previousOperationFuture
-                    .handleAsync((v, e) -> stopOperation.get(), replicaStartStopPool)
-                    .thenCompose(stopOperationFuture -> stopOperationFuture.thenApply(v -> {
-                        synchronized (context) {
-                            context.replicaState = ReplicaState.STOPPED;
-                        }
-
-                        LOG.debug("Weak replica stop complete [grpId={}, state={}].", groupId, context.replicaState);
-
-                        return true;
-                    }))
-                    .exceptionally(e -> {
-                        LOG.error("Replica stop failed [groupId={}]", e, groupId);
-
-                        throw new CompletionException(e);
-                    });
-
-            return context.previousOperationFuture.thenApply(v -> null);
-        }
-
-        private CompletableFuture<Void> planDeferredReplicaStop(
-                ReplicationGroupId groupId,
-                ReplicaStateContext context,
-                @Nullable HybridTimestamp leaseExpirationTime,
-                Supplier<CompletableFuture<Void>> deferredStopOperation
-        ) {
-            synchronized (context) {
-                // TODO IGNITE-23702: proper sync with waiting of expiration event, and proper deferred stop after cancellation of
-                //     reservation made by a lease that was not negotiated.
-                context.deferredStopReadyFuture = leaseExpirationTime == null
-                        ? new CompletableFuture<>()
-                        : replicaManager.clockService.waitFor(leaseExpirationTime);
-
-                return context.deferredStopReadyFuture
-                        .thenComposeAsync(unused -> stopReplica(groupId, context, deferredStopOperation), replicaManager.requestsExecutor);
-            }
-        }
-
-        private static void executeDeferredReplicaStop(ReplicationGroupId groupId, ReplicaStateContext context) {
-            assert context.deferredStopReadyFuture != null : "Stop operation future is not set [groupId=" + groupId + "].";
-
-            context.deferredStopReadyFuture.complete(null);
-            context.deferredStopReadyFuture = null;
-        }
-
-        /**
-         * Reserve replica as primary.
-         *
-         * @param groupId Group id.
-         * @return Whether the replica was successfully reserved.
-         */
-        boolean reserveReplica(ReplicationGroupId groupId, HybridTimestamp leaseStartTime) {
-            ReplicaStateContext context = getContext(groupId);
-
-            synchronized (context) {
-                ReplicaState state = context.replicaState;
-
-                if (state == ReplicaState.STOPPING || state == ReplicaState.STOPPED) {
-                    if (state == ReplicaState.STOPPED) {
-                        replicaContexts.remove(groupId);
-                    }
-
-                    if (context.reservedForPrimary) {
-                        throw new AssertionError("Unexpected replica reservation with " + state + " state [groupId=" + groupId + "].");
-                    }
-                } else if (state != ReplicaState.RESTART_PLANNED) {
-                    context.reserve(groupId, leaseStartTime);
-                }
-
-                return context.reservedForPrimary;
-            }
-        }
-
-        @TestOnly
-        boolean isReplicaPrimaryOnly(ReplicationGroupId groupId) {
-            ReplicaStateContext context = getContext(groupId);
-
-            synchronized (context) {
-                return context.replicaState == ReplicaState.PRIMARY_ONLY;
-            }
-        }
-    }
-
     /**
      * Destroys replication protocol storages for the given group ID.
      *
@@ -1570,137 +1219,6 @@ public class ReplicaManager extends AbstractEventProducer<LocalReplicaEvent, Loc
         RaftGroupOptions groupOptions = groupOptionsForPartition(isVolatileStorage, null);
 
         ((Loza) raftManager).destroyRaftNodeStorages(raftNodeId, groupOptions);
-    }
-
-    private static class ReplicaStateContext {
-        /** Replica state. */
-        ReplicaState replicaState;
-
-        /**
-         * Future of the previous operation, to linearize the starts and stops of replica. The result of the future is whether
-         * the operation was actually performed (for example, partition start operation can not start replica or raft node locally).
-         */
-        CompletableFuture<Boolean> previousOperationFuture;
-
-        /**
-         * Whether the replica is reserved to serve as a primary even if it is not included into assignments. If it is {@code} true,
-         * then {@link #weakStopReplica(ReplicationGroupId, WeakReplicaStopReason, Supplier)} transfers {@link ReplicaState#ASSIGNED}
-         * to {@link ReplicaState#PRIMARY_ONLY} instead of {@link ReplicaState#STOPPING}.
-         * Replica is reserved when it is in progress of lease negotiation and stays reserved when it's primary. The negotiation moves
-         * this flag to {@code true}. Primary replica expiration or the election of different node as a leaseholder moves this flag
-         * to {@code false}.
-         */
-        boolean reservedForPrimary;
-
-        /**
-         * Lease start time of the lease this replica is reserved for, not {@code null} if {@link #reservedForPrimary} is {@code true}.
-         */
-        @Nullable
-        HybridTimestamp leaseStartTime;
-
-        /**
-         * Future that should be complete when the deferred stop operation is ready to begin. Deferred stop operation is the stop of
-         * replica that was reserved for becoming primary, and needs to be stopped.
-         */
-        @Nullable
-        CompletableFuture<Void> deferredStopReadyFuture;
-
-        ReplicaStateContext(ReplicaState replicaState, CompletableFuture<Boolean> previousOperationFuture) {
-            this.replicaState = replicaState;
-            this.previousOperationFuture = previousOperationFuture;
-        }
-
-        void reserve(ReplicationGroupId groupId, HybridTimestamp leaseStartTime) {
-            if (reservedForPrimary && this.leaseStartTime != null && leaseStartTime.compareTo(this.leaseStartTime) < 0) {
-                // Newer lease may reserve this replica when it's already reserved by older one: this means than older one is no longer
-                // valid and most likely has not been negotiated and is discarded. By the same reason we shouldn't process the attempt
-                // of reservation by older lease, which is not likely and means reordering of message handling.
-                throw new IllegalArgumentException(format("Replica reservation failed: newer lease has already reserved this replica ["
-                        + "groupId={}, requestedLeaseStartTime={}, newerLeaseStartTime={}].", groupId, leaseStartTime,
-                        this.leaseStartTime));
-            }
-            this.leaseStartTime = leaseStartTime;
-            reservedForPrimary = true;
-        }
-
-        void unreserve() {
-            // TODO IGNITE-23702: should also lead to replica stop if it is PRIMARY_ONLY.
-            reservedForPrimary = false;
-            leaseStartTime = null;
-        }
-
-        void assertReservation(ReplicationGroupId groupId, HybridTimestamp leaseStartTime) {
-            assert reservedForPrimary : "Replica is elected as primary but not reserved [groupId="
-                    + groupId + ", leaseStartTime=" + leaseStartTime + "].";
-            assert leaseStartTime != null : "Replica is reserved but lease start time is null [groupId="
-                    + groupId + ", leaseStartTime=" + leaseStartTime + "].";
-        }
-    }
-
-    /**
-     * Replica lifecycle states.
-     * <br>
-     * Transitions:
-     * <br>
-     * On {@link #weakStartReplica(ReplicationGroupId, Supplier, Assignments)} (this assumes that the replica is included into assignments):
-     * <ul>
-     *     <li>if {@link #ASSIGNED}: next state is {@link #ASSIGNED};</li>
-     *     <li>if {@link #PRIMARY_ONLY}: next state is {@link #ASSIGNED};</li>
-     *     <li>if {@link #STOPPED} or {@link #STOPPING}: next state is {@link #STARTING}, replica is started after stop operation
-     *         completes;</li>
- *         <li>if {@link #RESTART_PLANNED}: produces {@link AssertionError} because replica should be stopped first;</li>
-     *     <li>if {@link #STARTING}: produces {@link AssertionError}.</li>
-     * </ul>
-     * On {@link #weakStopReplica(ReplicationGroupId, WeakReplicaStopReason, Supplier)} the next state also depends on given
-     * {@link WeakReplicaStopReason}:
-     * <ul>
-     *     <li>if {@link WeakReplicaStopReason#EXCLUDED_FROM_ASSIGNMENTS}:</li>
-     *     <ul>
-     *         <li>if {@link #ASSIGNED}: when {@link ReplicaStateContext#reservedForPrimary} is {@code true} then the next state
-     *             is {@link #PRIMARY_ONLY}, otherwise the replica is stopped, the next state is {@link #STOPPING};</li>
-     *         <li>if {@link #PRIMARY_ONLY} or {@link #STOPPING}: no-op.</li>
-     *         <li>if {@link #RESTART_PLANNED} no-op, because replica will be stopped within deferred operation;</li>
-     *         <li>if {@link #STARTING}: replica is stopped, the next state is {@link #STOPPING};</li>
-     *         <li>if {@link #STOPPED}: replica is stopped.</li>
-     *     </ul>
-     *     <li>if {@link WeakReplicaStopReason#PRIMARY_EXPIRED}:</li>
-     *     <ul>
-     *         <li>if {@link #PRIMARY_ONLY} replica is stopped, the next state is {@link #STOPPING}. Otherwise no-op.</li>
- *         </ul>
- *         <li>if {@link WeakReplicaStopReason#RESTART}: this is explicit manual replica restart for disaster recovery purposes,
-     *         replica is stopped. It happens immediately if it's <b>not</b> reserved as primary, the next state is {@link #STOPPING}. But
-     *         if if is reserved as primary, it asks the lease placement driver to stop the prolongation of lease, and is transferred
-     *         to the state {@link #RESTART_PLANNED}. When the lease is expired, the replica is stopped and transferred to
-     *         {@link #STOPPING} state.</li>
-     * </ul>
-     */
-    private enum ReplicaState {
-        /** Replica is starting. */
-        STARTING,
-
-        /**
-         * Local node, where the replica is located, is included into the union of stable and pending assignments. The replica can
-         * be either primary or non-primary. Assumes that the replica is started.
-         */
-        ASSIGNED,
-
-        /**
-         * Local node is excluded from the union of stable and pending assignments but the replica is a primary replica and hence
-         * can't be stopped. Assumes that the replica is started.
-         */
-        PRIMARY_ONLY,
-
-        /**
-         * Replica is going to be restarted, this state means that it is primary and needs to wait for lease expiration first.
-         * After lease is expired, replica is stopped and transferred to {@link ReplicaState#STOPPING} state.
-         */
-        RESTART_PLANNED,
-
-        /** Replica is stopping. */
-        STOPPING,
-
-        /** Replica is stopped. */
-        STOPPED
     }
 
     /**
